+++ contracts/oracle/OracleFactory.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.0;
 
-import '../interfaces/IOracleFactory.sol';
-import './Oracle.sol';
+import "../interfaces/IOracleFactory.sol";
+import "./Oracle.sol";
 
 // Used for constructing, updating, and consulting Oracle contracts
 contract OracleFactory is IOracleFactory {
@@ -29,8 +29,8 @@ contract OracleFactory is IOracleFactory {
      * @dev Create a new oracle for `token0` and `token1`.
      */
     function create(address token0, address token1) external {
-        require(oracles[token0][token1] == address(0), 'OracleFactory: ORACLE HAS ALREADY BEEN CREATED');
-        require(msg.sender == factory, 'OracleFactory: CALLER IS NOT FACTORY');
+        require(oracles[token0][token1] == address(0), "OracleFactory: ORACLE HAS ALREADY BEEN CREATED");
+        require(msg.sender == factory, "OracleFactory: CALLER IS NOT FACTORY");
 
         Oracle oracle = new Oracle(factory, token0, token1);
         oracles[token0][token1] = address(oracle);
@@ -43,7 +43,7 @@ contract OracleFactory is IOracleFactory {
      * @dev Update current prices of `token0` and `token1`.
      */
     function update(address token0, address token1) public {
-        require(oracles[token0][token1] != address(0), 'OracleFactory: ORACLE HAS NOT BEEN CREATED');
+        require(oracles[token0][token1] != address(0), "OracleFactory: ORACLE HAS NOT BEEN CREATED");
 
         Oracle oracle = Oracle(oracles[token0][token1]);
         if (canUpdate(oracle)) {
@@ -56,7 +56,7 @@ contract OracleFactory is IOracleFactory {
      * @dev Consult to get the `amountOut` of `tokenOut` equivalent in value to `amountIn` of `tokenIn`.
      *      Note this will always return 0 before update has been called successfully for the first time.
      */
-    function consult(address tokenIn, uint amountIn, address tokenOut) external view returns (uint amountOut) {
+    function consult(address tokenIn, uint256 amountIn, address tokenOut) external view returns (uint256 amountOut) {
         if (oracles[tokenIn][tokenOut] != address(0)) {
             Oracle oracle = Oracle(oracles[tokenIn][tokenOut]);
             amountOut = oracle.consult(tokenIn, amountIn);
